+++ vol/src/main/java/com/itu/vol/repository/TypeSiegeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TypeSiegeRepository extends JpaRepository<TypeSiege, Long> {
+public interface TypeSiegeRepository extends JpaRepository<TypeSiege, Integer> {
 }
